requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
ale-py=0.8.1=pypi_0
autorom=0.6.1=pypi_0
autorom-accept-rom-license=0.6.1=pypi_0
blas=1.0=mkl
box2d-py=2.3.5=pypi_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
ca-certificates=2024.3.11=haa95532_0
certifi=2024.6.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.2.0=py39h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
cython=3.0.10=pypi_0
decorator=4.4.2=pypi_0
docker-pycreds=0.4.0=pypi_0
durationpy=0.6=pypi_0
exceptiongroup=1.2.1=pypi_0
farama-notifications=0.0.4=pypi_0
filelock=3.15.4=pypi_0
fonttools=4.51.0=py39h2bbff1b_0
freetype=2.12.1=ha860e81_0
fsspec=2024.6.1=pypi_0
future=1.0.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
grpcio=1.64.1=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
gymnasium=0.29.1=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.7=pypi_0
imageio=2.34.1=pypi_0
imageio-ffmpeg=0.5.1=pypi_0
importlib-metadata=8.0.0=pypi_0
importlib_resources=6.1.1=py39haa95532_1
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=pypi_0
iso8601=2.1.0=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py39hd77b12b_0
krb5=1.20.1=h5b6d351_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libdeflate=1.17=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libtiff=4.5.1=hd77b12b_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_1
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.4=py39haa95532_0
matplotlib-base=3.8.4=py39h4ed8f06_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=pypi_0
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
moviepy=1.0.3=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.2.1=pypi_0
numpy=1.26.4=py39h055cbcc_0
numpy-base=1.26.4=py39h65a83cf_0
opencv-python=4.10.0.82=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_2
packaging=23.2=py39haa95532_0
paho-mqtt=2.1.0=pypi_0
pandas=2.2.2=pypi_0
pillow=10.3.0=py39h2bbff1b_0
pip=24.0=py39haa95532_0
platformdirs=4.2.2=pypi_0
pluggy=1.5.0=pypi_0
ply=3.11=py39haa95532_0
proglog=0.1.10=pypi_0
protobuf=4.25.3=pypi_0
psutil=6.0.0=pypi_0
pybind11-abi=5=hd3eb1b0_0
pygame=2.1.0=pypi_0
pygments=2.18.0=pypi_0
pyopengl=3.1.5=pypi_0
pyopengl-accelerate=3.1.5=pypi_0
pyparsing=3.0.9=py39haa95532_0
pyqt=5.15.10=py39hd77b12b_0
pyqt5-sip=12.13.0=py39h2bbff1b_0
pyserial=3.5=pypi_0
pytest=8.2.2=pypi_0
python=3.9.19=h1aa4202_1
python-dateutil=2.9.0post0=py39haa95532_2
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzed=4.1=pypi_0
qt-main=5.15.2=h879a1e9_9
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.13.1=py39h8640f81_0
sentry-sdk=2.7.1=pypi_0
serial=0.0.97=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=69.5.1=py39haa95532_0
shimmy=1.3.0=pypi_0
sip=6.7.12=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.1=pypi_0
sqlite=3.45.3=h2bbff1b_0
stable-baselines3=2.3.2=pypi_0
swig=4.2.1=pypi_0
sympy=1.12.1=pypi_0
tbb=2021.13.0=pypi_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tomli=2.0.1=py39haa95532_0
torch=2.3.1+cu118=pypi_0
torchaudio=2.3.1+cu118=pypi_0
torchvision=0.18.1+cu118=pypi_0
tornado=6.3.3=py39h2bbff1b_0
tqdm=4.66.4=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
unicodedata2=15.1.0=py39h2bbff1b_0
urllib3=2.2.1=pypi_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wandb=0.17.3=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39haa95532_0
xz=5.4.6=h8cc25b3_1
zipp=3.17.0=py39haa95532_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
